// angular-loja/src/app/fornecedores/models/fornecedorPost.ts
import { Endereco } from "src/app/shared/models/endereco";

export class FornecedorPost {

  idFornecedor?: number;
	nomeFornecedor: string;
  cpfCnpj: string;
  telefone1: string;
	telefone2: string;
	email: string;
	logradouro: string;
  numero: string;
  complemento: string;
  condominio: string;
  bairro: string;
  municipio: string;
  estado: string;
  cep: string;
}